Include ScreenScraper metadata reader unconditionally
Add an unconditional metadata-only ScreenScraper translation unit and APIs so menus can read local metadata sidecars without networking. Create network/screenscraper_meta.c and declare screenscraper_extended_media_allowed in the header; move metadata and rating helpers out of the network-only file. Remove HAVE_NETWORKING guards in menu code so scraped sidecar information is shown on builds without network support. Add the object to Makefile.common and include it in griffin.c.
